requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
sqlalchemy==2.0.40
psycopg2-binary==2.9.10
python-dotenv==1.1.0
pydantic==2.11.1
python-jose[cryptography]==3.3.0
bcrypt==4.2.0
alembic==1.13.1
apscheduler==3.10.4
python-multipart==0.0.20
anthropic>=0.40.0
pypdf>=4.0.0
Pillow>=10.0.0
google-api-python-client>=2.193.0
google-auth-oauthlib>=1.3.1
reportlab>=4.0.0
email-validator==2.3.0
slowapi==0.1.9
pytest==8.3.5
pytest-asyncio==0.24.0
httpx==0.28.1
